Add support for rail line overlays
This is dependent on specially-crafted signs placed in sequence.

All signs that are part of the same railway line must have the top
three lines of the front of the sign identical. Line 1 must be
'#[RAIL]', and lines 2+3 form the name of the line.

The last line of the front of the sign is four values separated by
slashes (`/`). The first sign in the sequence (ID 0) should have a
hex colour code in the first position. All other signs must have a
numeric sequence number in the first position. This need not be
monotonic nor positive, but it does need to be in the correct order
and unique. The remaining three values are grid offsets for the
location of the rail waypoints. If left blank, these default to zero.
For example, the bottom line might be `50///` or `80/0/1/0`.

I've left the old railway filter in place until we migrate in-game.

Depends-On: Iabbbc0da89b66beaa3967e211e57edcabb40148b
Change-Id: I4f0e134804d6f936877e4dcbf907b460a2b7cbd2