Provide player avatars locally
The old upstream avatar renderer seems to have been shut down.

I've got (experimental!) code to render the avatars, so I've done so
based on the current player markers that Overviewer generates. This is
a static, one-time-only thing until we figure out how we want to proceed
with player markers (be that continue with a static render, or move to
a service provided by the new upstream Overviewer)

Change-Id: Ice89fddfc1b7a910652405730709314a9f9ed340