Fix station markers not rendering station names
As the day and night renders share both a regionset and a set of
markers, the filters for each render are run twice on each POI, once for
each render. As the POI is passed in by reference, modifications to the
`poi` parameter in one filter function actually affect all filter
functions that touch this POI.

We *can* modify the POI in a filter function, but we need to be sure
that modifications still allow a filter function to work if the filter
function is called a second time on a POI. In other words, the filter
functions need to be idempotent to re-entry.

The only time we actually touch the marker in the rail functions is when
we are changing the side of the sign that the display text is on in
order to render it as a sign. In order to *not* affect the POI on other
renders, we need to modify a copy of the POI for the filter function.

We also need to sync the entire `x_text` element, not just the text or
things like colour and glow state won't be copied from the back.

Change-Id: I8a61afc3e2e847e8eb9b64033c243a68d1c0aa71